+++ x/common/denoms/denoms.go
@@ -1,9 +1,21 @@
 package denoms
 
+// Post Nibiru v2.6.0
+const (
+	// NOTE: US dollars. Use `denoms.USD` instead of `denoms.UUSD` going forward.
+	USD = "usd"
+	// Avalon Finance overcollateralized stablecoin.
+	// https://github.com/NibiruChain/pricefeeder/pull/64
+	USDA  = "usda"
+	SUSDA = "susda"
+)
+
+// Legacy denoms - These each include an unnecessary "u" prefix for micro.
+
 const ( // stablecoins
 	USDC = "uusdc"
 	NUSD = "unusd"
-	USD  = "uusd"
+	UUSD = "uusd"
 	USDT = "uusdt"
 )